Remove es_MX and pt_PT translations
As discussed with translators, it's more time-efficient to focus on the
"international" Spanish translation (`es`) and bring it near 100%
completion before branching it off for regional variants.
(So `es_MX` can be re-added copied from the near-complete `es` and
edited, instead of starting from scratch.)

Similarly for `pt_PT`, where focus should be put on finalizing the
`pt_BR` translation before branching it off and adapting it to the
European variant. (The current `pt_PT` translation was actually forked
from `pt_BR` already when it was around 25% complete, then did not see
much further work.)